Add files via upload
The primary challenges with the IoT based environments is security. It is essential to secure computer and Internet of Things platforms against a variety of threats and attacks. Furthermore, traditional intrusion detection systems (IDS) frequently handle massive amounts of data from IoT networks that contain redundant and irrelevant data, which results in decreased accuracy and longer response times. Therefore, the work proposed an IDS to detect different types of attacks on IoT networks by employing the CIC-IoT-2023 dataset.